+++ ui/src/app/shared/services/archived-workflows-service.ts
@@ -3,9 +3,9 @@ import {Pagination} from '../pagination';
 import requests from './requests';
 
 export class ArchivedWorkflowsService {
-    public list(namespace: string, name: string, phases: string[], labels: string[], minStartedAt: Date, maxStartedAt: Date, pagination: Pagination) {
+    public list(namespace: string, name: string, namePrefix: string, phases: string[], labels: string[], minStartedAt: Date, maxStartedAt: Date, pagination: Pagination) {
         return requests
-            .get(`api/v1/archived-workflows?${this.queryParams({namespace, name, phases, labels, minStartedAt, maxStartedAt, pagination}).join('&')}`)
+            .get(`api/v1/archived-workflows?${this.queryParams({namespace, name, namePrefix, phases, labels, minStartedAt, maxStartedAt, pagination}).join('&')}`)
             .then(res => res.body as models.WorkflowList);
     }
 
@@ -17,9 +17,18 @@ export class ArchivedWorkflowsService {
         return requests.delete(`api/v1/archived-workflows/${uid}`);
     }
 
+    public listLabelKeys() {
+        return requests.get(`api/v1/archived-workflows-label-keys`).then(res => res.body as models.Labels);
+    }
+
+    public listLabelValues(key: string) {
+        return requests.get(`api/v1/archived-workflows-label-values?listOptions.labelSelector=${key}`).then(res => res.body as models.Labels);
+    }
+
     private queryParams(filter: {
         namespace?: string;
         name?: string;
+        namePrefix?: string;
         phases?: Array<string>;
         labels?: Array<string>;
         minStartedAt?: Date;
@@ -40,6 +49,9 @@ export class ArchivedWorkflowsService {
         }
         if (filter.pagination.limit) {
             queryParams.push(`listOptions.limit=${filter.pagination.limit}`);
+        }
+        if (filter.namePrefix) {
+            queryParams.push(`namePrefix=${filter.namePrefix}`);
         }
         return queryParams;
     }