+++ workflow/signal/signal.go
@@ -6,9 +6,8 @@ import (
 	"strings"
 	"syscall"
 
-	corev1 "k8s.io/api/core/v1"
-
 	log "github.com/sirupsen/logrus"
+	corev1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/rest"
 
 	"github.com/argoproj/argo-workflows/v3/workflow/common"